REF: Move qt_kicker into separate module.
