+++ bluesky/callbacks.py
@@ -3,12 +3,9 @@ Useful callbacks for the Run Engine
 """
 import sys
 from itertools import count
-import asyncio
 import warnings
 from prettytable import PrettyTable
 
-import matplotlib.backends.backend_qt5
-from matplotlib.backends.backend_qt5 import _create_qApp
 import matplotlib.pyplot as plt
 
 from datetime import datetime
@@ -16,33 +13,6 @@ import numpy as np
 
 import logging
 logger = logging.getLogger(__name__)
-
-
-_create_qApp()
-qApp = matplotlib.backends.backend_qt5.qApp
-
-from matplotlib._pylab_helpers import Gcf
-
-try:
-    _draw_all = Gcf.draw_all  # mpl version >= 1.5
-except AttributeError:
-    # slower, but backward-compatible
-    def _draw_all():
-        for f_mgr in Gcf.get_all_fig_managers():
-            f_mgr.canvas.draw_idle()
-
-
-def _qt_kicker():
-    # The RunEngine Event Loop interferes with the qt event loop. Here we
-    # kick it to keep it going.
-    _draw_all()
-
-    qApp.processEvents()
-    loop.call_later(0.1, _qt_kicker)
-
-
-loop = asyncio.get_event_loop()
-loop.call_soon(_qt_kicker)
 
 
 class CallbackBase(object):
